requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-openai
langchain-pinecone
openai
pinecone-client
python-dotenv
tiktoken